Add a syslog logger implementation, using the `syslog` and `slog-syslog` crates. Warning: Doesn't work correctly.
Unfortunately, the `syslog` crate does not use the correct format for sending messages over a Unix-domain socket (`/dev/log` or `/var/run/log`), and the author is not responding to pull requests, so this is unlikely to ever be fixed. Also, the OpenBSD kernel has a dedicated system call for submitting syslog messages, so *any* pure-Rust syslog implementation will be sub-optimal on that platform.

For these reasons, I will next try to implement syslog for slog by calling into libc.